fix: correct paths to edit addresses
