+++ packages/default-theme/pages/_lang/account/addresses.vue
@@ -1,72 +1,9 @@
 <template>
-  <div class="shipping-list">
-      <SfTabs
-        key="address-list"
-        v-if="listAddresses"
-      >
-        <SfTab title="Shipping details">
-          <p class="message">
-            Manage all the shipping addresses you want (work place, home address
-            ...) This way you won't have to enter the shipping address manually
-            with each order.
-          </p>
-          <sw-address-list @editAddress="editAddress" />
-          <SwButton class="action-button" @click="changeAddress">
-            Add new address
-          </SwButton>
-        </SfTab>
-      </SfTabs>
-      <SfTabs v-else>
-        <SfTab title="Add address">
-          <nuxt-child />
-        </SfTab>
-      </SfTabs>
-  </div>
+  <nuxt-child />
 </template>
-<script>
-import { SfTabs } from '@storefront-ui/vue'
-import { useUser } from '@shopware-pwa/composables'
-import SwAddressList from '@shopware-pwa/default-theme/components/SwAddressList.vue'
-import SwButton from '@shopware-pwa/default-theme/components/atoms/SwButton'
 
+<script>
 export default {
-  name: 'MyAddresses',
-  components: {
-    SwButton,
-    SfTabs,
-    SwAddressList
-  },
-  data() {
-    return {
-      listAddresses: true
-    }
-  },
-  methods: {
-    changeAddress() {
-      this.$router.push(this.$i18n.path('addresses/add'))
-    },
-    editAddress(addressId) {
-      this.$router.push(this.$i18n.path(`addresses/add/${addressId}`))
-    },
-  },
-  watch: {
-    $route: {
-      deep: true,
-      handler(from, to) { 
-        this.listAddresses = !this.listAddresses
-      }
-    }
-  }
+  name: 'UserAddresses',
 }
 </script>
-
-<style lang="scss" scoped>
-@import '@/assets/scss/variables';
-
-
-
-.shipping-list {
-  margin-bottom: var(--spacer-xl);
-  width: 100%;
-}
-</style>
