+++ packages/default-theme/pages/_lang/account/addresses/add/_id.vue
@@ -1,20 +1,55 @@
 <template>
   <div class="addresses-add">
-    <SwAddress />
+    <div v-if="address !== ''">
+      <SwAddress :address="address" />
+    </div>
   </div>
 </template>
 
 <script>
 import SwAddress from '@shopware-pwa/default-theme/components/forms/SwAddress.vue'
+import { useUser } from '@shopware-pwa/composables'
 
 export default {
   components: { SwAddress },
   data() {
-    return {}
+    return {
+      address: '',
+    }
   },
-  mounted() {
-    console.log(this.$router.params.id)
-  }
+  setup() {
+    const {
+      addresses,
+      loadAddresses,
+      country,
+      loadCountry,
+      salutation,
+      loadSalutation,
+    } = useUser()
+    return {
+      loadAddresses,
+      addresses,
+      loadCountry,
+      country,
+      loadSalutation,
+      salutation,
+    }
+  },
+  async mounted() {
+    await this.loadAddresses()
+    const paramsId = this.$route.params && this.$route.params.id
+    if (paramsId) {
+      const address = this.addresses.find((addr) => addr.id === paramsId)
+      await this.loadCountry(address.countryId)
+      await this.loadSalutation(address.salutationId)
+
+      this.address = {
+        ...address,
+        country: this.country && this.country.data,
+        salutation: this.salutation && this.salutation.data,
+      }
+    }
+  },
 }
 </script>
