+++ packages/default-theme/pages/_lang/account/addresses/add/index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="addresses-add">
+    <SwAddress />
+  </div>
+</template>
+
+<script>
+import SwAddress from '@shopware-pwa/default-theme/components/forms/SwAddress.vue'
+
+export default {
+  components: { SwAddress },
+  data() {
+    return {}
+  }
+}
+</script>
+
+<style lang="scss" scoped></style>
