--- packages/default-theme/pages/account/addresses/add.vue
@@ -1,4 +1,0 @@
-<script>
-import Add from '@shopware-pwa/default-theme/pages/_lang/account/addresses/add'
-export default Add
-</script>
