+++ packages/default-theme/pages/account/addresses/add/index.vue
@@ -1,0 +1,4 @@
+<script>
+import Add from '@shopware-pwa/default-theme/pages/_lang/account/addresses/add/index'
+export default Add
+</script>
