+++ packages/default-theme/pages/account/addresses/index.vue
@@ -1,0 +1,4 @@
+<script>
+import Addresses from '@shopware-pwa/default-theme/pages/_lang/account/addresses/index'
+export default Addresses
+</script>
